Semana 4
El código de la semana 4, funcional

+++ TiendaBryan/src/main/java/Tienda/demo/domain/Categoria.java
@@ -25,7 +25,7 @@ public class Categoria implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_categoria")
-    private Long idCategoria;
+    private int idCategoria;
 
     @NotBlank(message = "La descripción no puede estar vacía")
     @Size(max = 50, message = "La descripción no puede exceder los 50 caracteres")
